Update nsenter.c
The function getopt_long returns a long integer, as opposed to char.  In "c" the char type is implementation depenedent. In X86/AMD  architecture char is treated as signed character and in Power PPC64le it is treated as unsigned char. It is better to downcast this to a char in the switch statement. Otherwise the while loop in Power architecture never terminates.